Roll the Dice Game
As a beginner Python coder, I am working on some mini projects to enhance my learning. Here is the first one.
It's a game in which the user rolls a dice up to 5 times, and gets the results and a total score calculated as the sum of all rolls.
Please leave feedback or tell me if you find some inputs that the user could make to break the code.