Add files via upload
This a simple trading strategy for the S&P Composite Index. I take monthly data from Quandl and apply a long-only trading strategy based on the 6MMA and the 24MMA. When the 6MMA is above the 12MMA the signal is to buy and when the 6MMA is below the 12MMA the signal is to sell.
Finally, I show the cumulative returns and sharp ratios to compare the performance of the index and the stragey.